Added experimental multi instance sweep, made more efficient device transfers.

+++ src/main/drl_from_sweep.py
@@ -8,45 +8,49 @@ from map import Map
 import wandb
 import json
 
-wandb.init(project="autonav")
+def start(delete_files=False):
+    wandb.init(project="autonav")
 
-LEN = 150000
+    LEN = 150000
 
-if sys.platform == "win32":
-    path = 'G:\\Projects\\AutoNav\\AutoNavServer\\assets\\testing\\BasicMap.json'
-    logger_path = "G:\Projects\AutoNav\AutoNavServer\output\logs"
-elif sys.platform == "linux" or sys.platform == "linux2":
-    path = '/home/jovyan/workspace/AutoNavServer/assets/testing/FRC2023Map.json'
-    wandb_path = '/home/jovyan/workspace/AutoNavServer/wandb'
-    logger_path = '/home/jovyan/workspace/AutoNavServer/output/logs'
-elif sys.platform == "darwin":
-    path = "/Users/maximkudryashov/Projects/AutoNav/AutoNavServer/assets/testing/BasicMap.json"
-    logger_path = "/Users/maximkudryashov/Projects/AutoNav/AutoNavServer/output/logs"
-else:
-    print("SYSTEM NOT SUPPORTED. EXITING")
-    exit()
-# path = '/Users/maximkudryashov/Projects/AutoNav/AutoNavServer/assets/testing/FRC2023Map.json'
+    if sys.platform == "win32":
+        path = 'G:\\Projects\\AutoNav\\AutoNavServer\\assets\\testing\\BasicMap.json'
+        logger_path = "G:\Projects\AutoNav\AutoNavServer\output\logs"
+    elif sys.platform == "linux" or sys.platform == "linux2":
+        path = '/home/jovyan/workspace/AutoNavServer/assets/testing/FRC2023Map.json'
+        wandb_path = '/home/jovyan/workspace/AutoNavServer/wandb'
+        logger_path = '/home/jovyan/workspace/AutoNavServer/output/logs'
+    elif sys.platform == "darwin":
+        path = "/Users/maximkudryashov/Projects/AutoNav/AutoNavServer/assets/testing/BasicMap.json"
+        logger_path = "/Users/maximkudryashov/Projects/AutoNav/AutoNavServer/output/logs"
+    else:
+        print("SYSTEM NOT SUPPORTED. EXITING")
+        exit()
+    # path = '/Users/maximkudryashov/Projects/AutoNav/AutoNavServer/assets/testing/FRC2023Map.json'
 
-JSON = json.load(open(path))
-mapobj = Map(JSON)
+    JSON = json.load(open(path))
+    mapobj = Map(JSON)
 
-#onshape_to_robot.
+    #onshape_to_robot.
 
-from_map(mapobj)
-DRL_VENV = RobotVEnv(map=mapobj, assets_path=ASSET_PATH)
+    from_map(mapobj)
+    DRL_VENV = RobotVEnv(map=mapobj, assets_path=ASSET_PATH)
 
 
-config = wandb.config
+    config = wandb.config
 
-TrainingExecutor = TrainingExecutor(mapobj, config=config, logger_path=logger_path)
+    TrainingExecutor = TrainingExecutor(mapobj, config=config, logger_path=logger_path)
 
-TrainingExecutor.train(DRL_VENV, config=config, total_ts=LEN, inpmap= mapobj, plotter_display=False)
-try:
-    for filename in os.listdir(wandb_path):
-        filepath = os.path.join(wandb_path, filename)
+    TrainingExecutor.train(DRL_VENV, config=config, total_ts=LEN, inpmap= mapobj, plotter_display=False)
+    if delete_files:
         try:
-            shutil.rmtree(filepath)
-        except OSError:
-            os.remove(filepath)
-except Exception as e:
-    print("Failed to remove wandb folders: " + str(e))
+            for filename in os.listdir(wandb_path):
+                filepath = os.path.join(wandb_path, filename)
+                try:
+                    shutil.rmtree(filepath)
+                except OSError:
+                    os.remove(filepath)
+        except Exception as e:
+            print("Failed to remove wandb folders: " + str(e))
+if __name__ == "__main__":
+    start(delete_files=True)
